# internal/handler/docs/appapi_swagger.yaml
basePath: /api/v1
definitions:
...
  auth.ForgotPasswordOption:
    properties:
      code:
        description: auth code get from user binding email
        example: F294484D
        type: string
      password:
        description: length of the password must be in range [10,30]
        example: "123456789"
        type: string
    type: object
...
  auth.LoginOption:
    properties:
      password:
        description: length of the password must be in range [10,30]
        example: "0123456789"
        type: string
      persistent:
        example: true
        type: boolean
      username:
        description: length of the username must be in range [6,20], and username
          must be Alphanumeric
        example: dstadmin
        type: string
    type: object
...
  dict.DictSaveOption:
    properties:
      code:
        type: string
      name:
        type: string
    type: object
...
  types.Response:
    properties:
      code:
        example: 2000
        type: integer
      data: {}
      err:
        type: string
      msg:
        example: operation success
        type: string
    type: object
...
info:
  contact:
    name: dstgo
    url: https://github.com/dstgo
  description: |-
    app http api documentation, use the Bearer Token to authenticate
    It should be noted that when using swagger doc for API debugging, the Token needs to be manually prefixed with Bearer.
  license:
    name: MIT LICENSE
    url: https://mit-license.org/
  title: App Internal API Documentation
  version: v1.0.0
paths:
  /auth/forgotpwd:
    post:
      consumes:
      - application/json
      description: |-
        [guest]
        forgot password
      parameters:
      - description: Account ID
        in: body
        name: form
        required: true
        schema:
          $ref: '#/definitions/auth.ForgotPasswordOption'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.Response'
      summary: ForgotPassword
      tags:
      - auth
  /auth/login:
    post:
      consumes:
      - application/json
      description: |-
        [guest]
        if login success, return jwt token
      parameters:
      - description: comment
        in: body
        name: loginBody
        required: true
        schema:
          $ref: '#/definitions/auth.LoginOption'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/types.Response'
            - properties:
                data:
                  $ref: '#/definitions/auth.Token'
              type: object
      summary: Login
      tags:
      - auth
  /auth/logout:
    delete:
      consumes:
      - application/json
      description: |-
        [user]
        user logout
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.Response'
      security:
      - BearerAuth: []
      summary: Logout
      tags:
      - auth
  /auth/refresh:
    get:
      consumes:
      - application/json
      description: |-
        [guest]
        carry refresh token in query params, access token in header
        if refresh-token expired , TokenRefresher will not refresh token [4012]
        else if access-token has expired after delay duration, TokenRefresher will not refresh token [4012]
        else if access-token has expired before delay duration, TokenRefresher will issue a new access-token [2005]
        else if access-token has not expired, TokenRefresher will renewal the 1/10 access-token ttl per time  [2005]
        else if access-token has not expired, and ttl >= 2 * conf.JwtConf.Exp, TokenRefresher will not refresh token [4013]
      parameters:
      - description: refresh token
        in: query
        name: refresh
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/types.Response'
            - properties:
                data:
                  $ref: '#/definitions/auth.Token'
              type: object
      summary: Refresh
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: |-
        [guest]
        user register api
      parameters:
      - description: comment
        in: body
        name: registerBody
        required: true
        schema:
          $ref: '#/definitions/auth.RegisterOption'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.Response'
      summary: Register
      tags:
      - auth
  /dict/create:
    post:
      consumes:
      - application/json
      description: create dict
      parameters:
      - description: DictSaveOption
        in: body
        name: DictSaveOption
        required: true
        schema:
          $ref: '#/definitions/dict.DictSaveOption'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.Response'
      security:
      - BearerAuth: []
      summary: CreateDict
      tags:
      - dict
  /dict/data/create:
    post:
      consumes:
      - application/json
      description: get string by ID
      parameters:
      - description: DictDataSaveOption
        in: body
        name: DictDataSaveOption
        required: true
        schema:
          $ref: '#/definitions/dict.DictDataSaveOption'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.Response'
      security:
      - BearerAuth: []
      summary: CreateDictData
      tags:
      - dict
  /dict/data/list:
    get:
      consumes:
      - application/json
      description: list dict data
      parameters:
      - in: query
        name: code
        type: string
      - description: if is reverse order
        example: true
        in: query
        name: desc
        type: boolean
      - description: page
        example: 1
        in: query
        name: page
        type: integer
      - in: query
        name: search
        type: string
      - description: size of page
        example: 10
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/types.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dict.DictDataDetail'
                  type: array
              type: object
      security:
      - BearerAuth: []
      summary: ListDictData
      tags:
      - dict
  /dict/data/remove:
    delete:
      consumes:
      - application/json
      description: get string by ID
      parameters:
      - in: query
        name: id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.Response'
      security:
      - BearerAuth: []
      summary: RemoveDictData
      tags:
      - dict
  /dict/data/update:
    post:
      consumes:
      - application/json
      description: get string by ID
      parameters:
      - description: DictDataUpdateOption
        in: body
        name: DictDataUpdateOption
        required: true
        schema:
          $ref: '#/definitions/dict.DictDataUpdateOption'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.Response'
      security:
      - BearerAuth: []
      summary: UpdateDictData
      tags:
      - dict
  /dict/info:
    get:
      consumes:
      - application/json
      description: get dict data info by code
      parameters:
      - in: query
        name: code
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/types.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dict.DictDataInfo'
                  type: array
              type: object
      security:
      - BearerAuth: []
      summary: GetDictInfo
      tags:
      - dict
  /dict/list:
    get:
      consumes:
      - application/json
      description: list dict pages
      parameters:
      - description: if is reverse order
        example: true
        in: query
        name: desc
        type: boolean
      - description: page
        example: 1
        in: query
        name: page
        type: integer
      - in: query
        name: search
        type: string
      - description: size of page
        example: 10
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/types.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dict.DictDetail'
                  type: array
              type: object
      security:
      - BearerAuth: []
      summary: ListDict
      tags:
      - dict
  /dict/remove:
    delete:
      consumes:
      - application/json
      description: remove dict
      parameters:
      - in: query
        name: id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.Response'
      security:
      - BearerAuth: []
      summary: RemoveDict
      tags:
      - dict
  /dict/update:
    post:
      consumes:
      - application/json
      description: update dict
      parameters:
      - description: DictUpdateOption
        in: body
        name: DictUpdateOption
        required: true
        schema:
          $ref: '#/definitions/dict.DictUpdateOption'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.Response'
      security:
      - BearerAuth: []
      summary: UpdateDict
      tags:
      - dict
...
  /key/create:
    post:
      consumes:
      - application/json
      description: create specified user api key
      parameters:
      - description: KeyCreateOption
        in: body
        name: KeyCreateOption
        required: true
        schema:
          $ref: '#/definitions/auth.KeyCreateOption'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.Response'
      security:
      - BearerAuth: []
      summary: CreateAPIKey
      tags:
      - key
  /key/list:
    get:
      consumes:
      - application/json
      description: list specified user api keys
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/types.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/auth.APIKey'
                  type: array
              type: object
      security:
      - BearerAuth: []
      summary: ListAPIKeys
      tags:
      - key
  /key/remove:
    delete:
      consumes:
      - application/json
      description: remove specified api key
      parameters:
      - in: query
        name: key
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.Response'
      security:
      - BearerAuth: []
      summary: RemoveAPIKey
      tags:
      - key
  /perm/create:
    post:
      consumes:
      - application/json
      description: |-
        [admin]
        create a new permission
      parameters:
      - description: create perm
        in: body
        name: createPerm
        required: true
        schema:
          $ref: '#/definitions/role.CreatePermOption'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.Response'
      security:
      - BearerAuth: []
      summary: CreatePermission
      tags:
      - role
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"